Removed the redundant 'name' field from the functions that get data for an individual concoction
The frontend code is making two concoction-related GET requests.
The first request is for all concoctions. It gets each concoction's id, name, and created_at field. This data is then saved on the frontend.
The second request is for an individual concoction. It gets additional data (i.e. instructions and notes).
Thus, the concoction's name does not need to be sent back a second time.