Created a function that retrieves a concoction's coffee
For now, a concoction will have only one coffee, as that should be the most common use case.
This can be changed later, but a user could theoretically add another coffee as an ingredient.
I also added an edge case that should never happen; a concoction should always have coffee!