VERY minor edit: Updated the seeds file and shortened the length of a coffee's roast name
On the frontend, I decided to give users the following roast options: Blonde, Light, Medium-Light, Medium, Medium-Dark, and Dark.
I have chosen not to include Blonde Espresso, as that is more of a roast AND blend.
As such, I can decrease the maximum length of the roast name in the database.