Moved the file with the handleServerError function into another folder, renamed the file, and used it to refactor authMiddleware.js
I realized that the handleServerError function could also be used by the functions in the authMiddleware file.
For that reason, I moved the controllerUtils file into a utils/ folder and renamed it errorStatuses.js.
I also moved some code outside of try-catch blocks, just like in the previous commit, and for the same reason.