Enabled users to delete their profiles
Note: in the userController's deleteProfile function, the edge case where it cannot find a user with the provided user id,
should not happen unless either something goes wrong with authentication or someone tries to use an expired user's session tokens.

I also added a note to the seeds file. Whoever uses it will need to create five user accounts in order for the SQL queries in that file to work correctly.